+++ src/main/java/de/relluem94/minecraft/server/spigot/essentials/model/CustomInventory.java
@@ -1,4 +1,4 @@
-package de.relluem94.minecraft.server.spigot.essentials.helpers.objects;
+package de.relluem94.minecraft.server.spigot.essentials.model;
 
 
 import de.relluem94.minecraft.server.spigot.essentials.helpers.ItemHelper;
