+++ src/main/java/de/relluem94/minecraft/server/spigot/essentials/model/enchantment/CustomEnchantment.java
@@ -1,12 +1,12 @@
-package de.relluem94.minecraft.server.spigot.essentials.enchantment;
+package de.relluem94.minecraft.server.spigot.essentials.model.enchantment;
 
 import com.google.common.collect.Multimap;
 import de.relluem94.minecraft.server.spigot.essentials.helpers.ItemHelper;
 import lombok.Getter;
 import org.bukkit.NamespacedKey;
-import org.bukkit.enchantments.EnchantmentTarget;
 import org.bukkit.attribute.Attribute;
 import org.bukkit.attribute.AttributeModifier;
+import org.bukkit.enchantments.EnchantmentTarget;
 
 public class CustomEnchantment {
     @Getter
