+++ src/test/java/de/relluem94/minecraft/server/spigot/essentials/model/pojo/CropEntryTest.java
@@ -1,4 +1,4 @@
-package de.relluem94.minecraft.server.spigot.essentials.helpers.pojo;
+package de.relluem94.minecraft.server.spigot.essentials.model.pojo;
 
 import org.bukkit.Material;
 import org.junit.jupiter.api.BeforeEach;
